requirements.txt
asgiref==3.9.2
Django==4.2.24
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0